refactor: move prune tokio logic to ethrex crate
